Move Message interface to the types
